backend/requirements.txt
fastapi
uvicorn
pydantic
pypdf
google-genai
python-dotenv
google-adk
